\name{computeCentroids}
\alias{computeCentroids}
\title{computeCentroids}
\usage{
  computeCentroids(X, idx, K)
}
\arguments{
  \item{X}{dataset}

  \item{idx}{cluster index}

  \item{K}{cluster number}
}
\value{
  centroids
}
\description{
  computing centroids
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}

