% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{computeCost}
\alias{computeCost}
\title{computeCost}
\usage{
computeCost(X, y, theta, lambda = 0)
}
\arguments{
\item{X}{x values (a column of 1 was added)}

\item{y}{y values}

\item{theta}{theta values}
}
\value{
cost J
}
\description{
computing the cost J
}
\details{
The gradient descent algorithm was to minimize the cost function.
}
\author{
Guangchuang Yu \url{https://yulab-smu.top}
}
