% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlass-package.R
\docType{data}
\name{DataSet}
\alias{DataSet}
\alias{ex1data1}
\alias{ex2data1}
\alias{ex6data1}
\alias{ex6data2}
\alias{ex7data2}
\title{Datasets
data samples.}
\description{
Datasets
data samples.
}
\keyword{datasets}
