\name{findClosestCentroids}
\alias{findClosestCentroids}
\title{findClosestCentroids}
\usage{
  findClosestCentroids(X, centroids)
}
\arguments{
  \item{X}{dataset}

  \item{centroids}{centroids}
}
\value{
  cluster index
}
\description{
  finding cloest centroids
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}

