% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hcluster}
\alias{hcluster}
\title{hcluster}
\usage{
hcluster(mat, method = "complete", dist.method = "euclidean")
}
\arguments{
\item{mat}{matrix}

\item{method}{one of single, complete, and average}

\item{dist.method}{method for calculating distance}
}
\value{
list
}
\description{
hiarachical cluster matrix
}
\author{
G Yu
}

