% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kMeans.R
\name{kMeans}
\alias{kMeans}
\title{kMeans}
\usage{
kMeans(X, centers, max.iter = 10, lang = "CPP")
}
\arguments{
\item{X}{dataset}

\item{centers}{initial centers or K}

\item{max.iter}{maximum iterations}

\item{lang}{R or CPP}
}
\value{
A \code{kMeansResult} instance.
}
\description{
kMeans algorithm
}
\details{
kmeans algorithm
}
\author{
Guangchuang Yu \url{http://ygc.name}
}
\keyword{manip}
