% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{kMeansResult-class}
\alias{"[",kMeansResult-method}
\alias{kMeansResult-class}
\alias{plot,kMeansResult-method}
\title{Class "kMeansResult"
This class represents the result of kmeans clustering.}
\description{
Class "kMeansResult"
This class represents the result of kmeans clustering.
}
\section{Slots}{

\describe{
\item{\code{dataset}}{original dataset}

\item{\code{clusters}}{clustering result}

\item{\code{centroids}}{final centroids}

\item{\code{K}}{K clusters}

\item{\code{traceCentroids}}{tracing the centroids when algorithm progress}
}}
\author{
Guangchuang Yu \url{http://ygc.name}
}
\seealso{
\code{\link{kMeans}}
}
\keyword{classes}

