% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logisticRegression}
\alias{logisticRegression}
\title{logisticRegression}
\usage{
logisticRegression(X, y, max.iter = 15, lambda = 0)
}
\arguments{
\item{X}{x values}

\item{y}{labels}

\item{max.iter}{number of iteration}
}
\value{
A \code{logisticRegressionResult} instance.
}
\description{
Logistic Regression algorithm
}
\author{
Yu Guangchuang \url{http://ygc.name}
}
\keyword{manip}

