% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{logisticRegressionResult-class}
\alias{"[",}
\alias{logisticRegressionResult-class}
\alias{logisticRegressionResult-method}
\alias{plot,}
\title{Class "logisticRegressionResult"
This class represents the result of logistic regression.}
\description{
Class "logisticRegressionResult"
This class represents the result of logistic regression.
}
\section{Slots}{

\describe{
\item{\code{X}}{x values (a column of 1 was added)}

\item{\code{y}}{y values}

\item{\code{theta}}{theta values}

\item{\code{lambda}}{parameter for regularization}
}}
\author{
Guangchuang Yu \url{http://ygc.name}
}
\seealso{
\code{\link{logisticRegression}}
}
\keyword{classes}

