\docType{package}
\name{mlass-package}
\alias{mlass}
\alias{mlass-package}
\title{ML algorithms implemented for ML class.}
\description{
  This package is designed to implement the ML algorithms
  presented in ML Class (http://www.ml-class.org,
  Stanford).
}
\details{
  \tabular{ll}{ Package: \tab mlass\cr Type: \tab
  Package\cr Version: \tab 0.3.0\cr Date: \tab
  06-20-2012\cr Depends:\tab ggplot2\cr License: \tab
  Artistic-2.0\cr }
}
\author{
  Guangchuang Yu \email{guangchuangyu@gmail.com}

  Maintainer: Guangchuang Yu
  \email{guangchuangyu@gmail.com}
}
\seealso{
  \linkS4class{kMeansResult}, \link{kMeans}
}
\keyword{package}

