\docType{package}
\name{MLclass-package}
\alias{MLclass}
\alias{MLclass-package}
\title{statistical analysis and visulization of functional profiles for genes and
gene clusters
The package implements methods to analyze and visualize functional profiles
of gene and gene clusters.}
\description{
  This package is designed to implement the ML algorithms
  presented in ML Class (http://www.ml-class.org,
  Stanford).
}
\details{
  \tabular{ll}{ Package: \tab MLclass\cr Type: \tab
  Package\cr Version: \tab 0.1.0\cr Date: \tab
  12-29-2011\cr Depends:\tab ggplot2, methods\cr License:
  \tab Artistic-2.0\cr }
}
\author{
  Guangchuang Yu \email{guangchuangyu@gmail.com}

  Maintainer: Guangchuang Yu
  \email{guangchuangyu@gmail.com}
}
\seealso{
  \linkS4class{kMeansResult}, \link{kMeans}
}
\keyword{package}

