% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{svmResult-class}
\alias{"[",svmResult-method}
\alias{plot,svmResult-method}
\alias{svmResult-class}
\title{Class "svmResult"
This class represents the result of svm training model.}
\description{
Class "svmResult"
This class represents the result of svm training model.
}
\section{Slots}{

\describe{
\item{\code{X}}{dataset}

\item{\code{y}}{label}

\item{\code{kernelFunction}}{kernnelFunction}

\item{\code{w}}{w}

\item{\code{b}}{b}

\item{\code{alphas}}{alphas}
}}
\author{
Guangchuang Yu \url{http://ygc.name}
}
\seealso{
\code{\link{svmTrain}}
}
\keyword{classes}

