% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcluster.R
\name{hcluster}
\alias{hcluster}
\title{hcluster}
\usage{
hcluster(mat, method = "complete", dist.method = "euclidean")
}
\arguments{
\item{mat}{matrix}

\item{method}{one of single, complete, and average}

\item{dist.method}{method for calculating distance}
}
\value{
list
}
\description{
hiarachical cluster matrix
}
\author{
G Yu
}
