\name{kMeans}
\alias{kMeans}
\title{kMeans}
\usage{
kMeans(X, centers, max.iter = 10)
}
\arguments{
  \item{X}{dataset}

  \item{centers}{initial centers or K}

  \item{max.iter}{maximum iterations}
}
\value{
  A \code{kMeansResult} instance.
}
\description{
  kMeans algorithm
}
\details{
  kmeans algorithm
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}
\keyword{manip}
