% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kMeans.R
\docType{class}
\name{kMeansResult-class}
\alias{kMeansResult-class}
\alias{"[",kMeansResult-method}
\alias{plot,kMeansResult-method}
\title{Class "kMeansResult"
This class represents the result of kmeans clustering.}
\description{
Class "kMeansResult"
This class represents the result of kmeans clustering.
}
\section{Slots}{

\describe{
\item{\code{dataset}}{original dataset}

\item{\code{clusters}}{clustering result}

\item{\code{centroids}}{final centroids}

\item{\code{K}}{K clusters}

\item{\code{traceCentroids}}{tracing the centroids when algorithm progress}
}}

\seealso{
\code{\link{kMeans}}
}
\author{
Guangchuang Yu \url{http://ygc.name}
}
\keyword{classes}
