\docType{class}
\name{kMeansResult-class}
\alias{getClusters,}
\alias{kMeansResult-class}
\alias{kMeansResult-method}
\alias{plot,}
\title{Class "kMeansResult"
This class represents the result of kmeans clustering.}
\description{
  Class "kMeansResult" This class represents the result of
  kmeans clustering.
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}
\seealso{
  \code{\link{kMeans}}
}
\keyword{classes}

