% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\docType{class}
\name{linearRegressionResult-class}
\alias{linearRegressionResult-class}
\alias{linearRegressionResult-method}
\title{Class "linearRegressionResult"
This class represents the result of linear regression.}
\description{
Class "linearRegressionResult"
This class represents the result of linear regression.
}
\section{Slots}{

\describe{
\item{\code{X}}{x values (a column of 1 was added)}

\item{\code{y}}{y values}

\item{\code{theta}}{theta values}

\item{\code{degree}}{feature degree}
}}

\author{
Guangchuang Yu \url{https://yulab-smu.top}
}
\keyword{classes}
