\docType{class}
\name{linearRegressionResult-class}
\alias{getTheta,}
\alias{linearRegressionResult-class}
\alias{linearRegressionResult-method}
\alias{lineareRegressionResult-method}
\alias{plot,}
\title{Class "linearRegressionResult"
This class represents the result of linear regression.}
\description{
  Class "linearRegressionResult" This class represents the
  result of linear regression.
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}
\seealso{
  \code{\link{gradDescent}}
}
\keyword{classes}

