\docType{class}
\name{logisticRegressionResult-class}
\alias{"[",}
\alias{logisticRegressionResult-class}
\alias{logisticRegressionResult-method}
\alias{plot,}
\title{Class "logisticRegressionResult"
This class represents the result of logistic regression.}
\description{
  Class "logisticRegressionResult" This class represents
  the result of logistic regression.
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}
\seealso{
  \code{\link{logisticRegression}}
}
\keyword{classes}

