% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlass-package.R
\docType{package}
\name{mlass-package}
\alias{mlass}
\alias{mlass-package}
\title{mlass: Machine Learning Algorithms}
\description{
This package is designed to implement the ML algorithms presented in ML Class (\url{http://www.ml-class.org}, Stanford).
}
\keyword{internal}
