% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,kMeansResult}
\alias{plot,linearRegressionResult,ANY-method}
\alias{plot,logisticRegressionResult,ANY-method}
\alias{plot,svmResult}
\title{plot method}
\usage{
\S4method{plot}{kMeansResult,ANY}(x, trace = F, title = "", xlab = "",
  ylab = "")

\S4method{plot}{linearRegressionResult,ANY}(x, title = "", xlab = "",
  ylab = "")

\S4method{plot}{logisticRegressionResult,ANY}(x, title = "", xlab = "",
  ylab = "")

\S4method{plot}{svmResult,ANY}(x, X, y, type, title = "", xlab = "",
  ylab = "")
}
\arguments{
\item{x}{A \code{kMeans} instance}

\item{trace}{tracing centroids when algorithm progress}

\item{object}{A \code{linearRegressionResult} instance.}

\item{object}{A \code{logisticRegressionResult} instance.}

\item{x}{A \code{svmResult} instance}

\item{X}{data matrix}

\item{y}{class label, 0 or 1}

\item{type}{one of linear or nonlinear}

\item{title}{title}

\item{xlab}{xlab}

\item{ylab}{ylab}
}
\value{
graph

ggplot object

ggplot object

ggplot2 graph object
}
\description{
plot method for \code{kMeansResult} instance

plot method for \code{linearRegressionResult} instance

plot method for \code{logisticRegressionResult} instance

plot method for \code{svmResult} instance
}
\author{
Guangchuang Yu \url{http://ygc.name}

Guangchuang Yu \url{http://ygc.name}

Guangchuang Yu \url{http://ygc.name}

Guangchuang Yu \url{http://ygc.name}
}

