\docType{methods}
\name{plot}
\alias{plot}
\title{plot method}
\arguments{
  \item{object}{A \code{kMeans} instance}

  \item{trace}{tracing centroids when algorithm progress}

  \item{object}{A \code{linearRegressionResult} instance.}
}
\value{
  graph

  ggplot object
}
\description{
  plot method for \code{kMeansResult} instance

  plot method for \code{linearRegressionResult} instance
}
\author{
  Guangchuang Yu \url{http://ygc.name}

  Guangchuang Yu \url{http://ygc.name}
}

