\docType{methods}
\name{plot}
\alias{plot}
\title{plot method}
\arguments{
  \item{object}{A \code{kMeans} instance}

  \item{trace}{tracing centroids when algorithm progress}
}
\value{
  graph
}
\description{
  plot method for \code{kMeansResult} instance
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}
