% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm.R
\docType{class}
\name{svmResult-class}
\alias{svmResult-class}
\alias{"[",svmResult-method}
\alias{plot,svmResult-method}
\title{Class "svmResult"
This class represents the result of svm training model.}
\description{
Class "svmResult"
This class represents the result of svm training model.
}
\section{Slots}{

\describe{
\item{\code{X}}{dataset}

\item{\code{y}}{label}

\item{\code{kernelFunction}}{kernnelFunction}

\item{\code{w}}{w}

\item{\code{b}}{b}

\item{\code{alphas}}{alphas}
}}

\seealso{
\code{\link{svmTrain}}
}
\author{
Guangchuang Yu \url{https://yulab-smu.top}
}
\keyword{classes}
