\docType{class}
\name{svmResult-class}
\alias{plot,svmResult-method}
\alias{svmResult-class}
\alias{"[",svmResult-method}
\title{Class "svmResult"
This class represents the result of svm training model.}
\description{
  Class "svmResult" This class represents the result of svm
  training model.
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}
\seealso{
  \code{\link{svmTrain}}
}
\keyword{classes}

